+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Unreleased
 
 -   Add new (experimental) list, and add depreciation notice for the current list. #6787
+-   Force `<SearchListItem>` form elements id to be unique. #6871
+-   Add `controlId` and `name` props to `<SearchListItem>`. #6871
+-   Minor styling tweaks and fixes to `<SearchListcontrol>`. #6871
 
 # 6.1.2